requirements.txt
Django==3.2.9
psycopg2-binary==2.9.3
djangorestframework==3.13.1
django-cors-headers==3.11.0
gunicorn==20.1.0
django-environ==0.8.1
drf-yasg2==1.19.4
coreapi==2.3.3
coreschema==0.0.4
djangorestframework-jsonapi==5.0.0
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
djantic==0.7.0
pyshorteners==1.0.1
